fixing utils/osutils_linux.go as recommended

+++ utils/osutils_linux.go
@@ -67,7 +67,7 @@ func GetFilesystemStats(
 	}
 
 	buf := result.Output
-	size := int64(buf.Blocks) * buf.Bsize
+	size := int64(buf.Blocks) * int64(buf.Bsize)
 	Logc(ctx).WithFields(log.Fields{
 		"path":   path,
 		"size":   size,
@@ -77,7 +77,7 @@ func GetFilesystemStats(
 		"free":   buf.Bfree,
 	}).Debug("Filesystem size information")
 
-	available = int64(buf.Bavail) * buf.Bsize
+	available = int64(buf.Bavail) * int64(buf.Bsize)
 	capacity = size
 	usage = capacity - available
 	inodes = int64(buf.Files)
